Create encrypted /srv bound with clevis to the TPM2
On reinstall work around anaconda and don't format /srv and /home.
Also reconstruct missing /etc/crypttab and /etc/fstab for /srv.

Install clevis systemd units to open a luks in local-fs.target.
It does not need network and remote-fs.
Also we want to have it ready before any other daemons start,
in case we have the configuration in /srv.